+++ src/turismo_en_la_tierra_media/PromoAbsoluta.java
@@ -2,13 +2,13 @@ package turismo_en_la_tierra_media;
 
 public class PromoAbsoluta extends Promocion {
 
-	public PromoAbsoluta(String nombreAtraccion, Double costoPromocion, Atraccion[] atracciones) {
-		super(nombreAtraccion, TipoPromocion.ABSOLUTA, 0.0, atracciones);
+	public PromoAbsoluta(String nombreAtraccion, Integer costoPromocion, Atraccion[] atracciones) {
+		super(nombreAtraccion, TipoPromocion.ABSOLUTA, 0, atracciones);
 		this.setCostoAtraccion(costoPromocion);
 	}
 
 	@Override
-	public void setCostoAtraccion(Double costoPromocion) {
+	public void setCostoAtraccion(Integer costoPromocion) {
 		super.costoPromocion = costoPromocion;
 	}
